+++ CHANGELOG.md
@@ -43,6 +43,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * ffi(sdk): add `sqlite` and `ndb` features ([Yuki Kishimoto])
 * js(nostr): add missing NIP51 constructors ([rustedmoon])
 * js(nostr): add NIP47 request params and response results structs ([Yuki Kishimoto])
+* js(sdk): add `NWC` client ([Yuki Kishimoto])
 
 ### Fixed